Feat: Added Close message type and Keyword Type enum
Close message type means the button type that is attached to the closing
message of a ticket. Possible types are CLOSE_TOGGLE, CLOSE, and
AFTER_CLOSE.

Possible keyword types are MATCH_START, IS_SUBSTR. The MATCH_START type
means a response is sent if a message starts with a keyword. The
IS_SUBSTR type means a response is sent whenever a keyword is a
substring of a message.